+++ commonFooter.php
@@ -1,4 +1,4 @@
-	<script type="module" src="/html/<?= VERSION ?>/js/lizyVideos.js"></script>
+	<script type="module" src="/html/<?= VERSION ?>/js/lazyVideos.js"></script>
 	<script type="module" src="/html/<?= VERSION ?>/js/modules/lazysizes.min.js"></script>
 	<script src="https://unpkg.com/flickity@2/dist/flickity.pkgd.min.js"></script>
 	<script src="https://unpkg.com/flickity-fade@1/flickity-fade.js"></script>
